Refactoring following rebase from master.
